+++ docs/CHANGE_LOG.md
@@ -3,6 +3,7 @@
 ## Version 0.5.0:
 
 #### **Public Functions:**
+- BREAKING: Renamed "json() > includeCssStyles()" to "json() > includeCssProperties()".
 - When "addCssToHead()" is used for "html()", all the CSS is now placed in one "style" node in the HTML document head, instead of multiple ones!
 - When "addCssToHead()" is used for "html()", the CSS will now use the CSS class from the DOM element (it reverts back to the DOM element ID if not available).
 - Added "ignoreNodeTypes()" support for "json()", which states the node types that should not be included in the JSON returned from "get()".